+++ app/src/main/kotlin/de/binarynoise/captiveportalautologin/json/webRequest/OnErrorOccurredDetails.kt
@@ -2,7 +2,7 @@ package de.binarynoise.captiveportalautologin.json.webRequest
 
 import kotlinx.serialization.json.JsonObject
 import de.binarynoise.util.json.getBoolean
-import de.binarynoise.util.json.getFloat
+import de.binarynoise.util.json.getDouble
 import de.binarynoise.util.json.getInt
 import de.binarynoise.util.json.getLong
 import de.binarynoise.util.json.getOptString
@@ -41,7 +41,7 @@ class OnErrorOccurredDetails(
     val documentUrl: String? = null,
     val tabId: Int,
     val type: String,
-    val timeStamp: Float,
+    val timeStamp: Double,
     val ip: String? = null,
     val fromCache: Boolean,
     val error: String,
@@ -58,7 +58,7 @@ class OnErrorOccurredDetails(
                 json.getOptString("documentUrl"),
                 json.getInt("tabId"),
                 json.getString("type"),
-                json.getFloat("timeStamp"),
+                json.getDouble("timeStamp"),
                 json.getOptString("ip"),
                 json.getBoolean("fromCache"),
                 json.getString("error"),
